+++ src/testing/context.ts
@@ -2,8 +2,12 @@ import {
   IntegrationInstance,
   IntegrationExecutionContext,
   IntegrationStepExecutionContext,
+  IntegrationInstanceConfigField,
+  IntegrationInstanceConfigFieldMap,
 } from '../framework';
 
+import { loadInstanceConfigFields } from '../framework/config';
+import { loadConfigFromEnvironmentVariables } from '../framework/execution/config';
 import { LOCAL_INTEGRATION_INSTANCE } from '../framework/execution/instance';
 
 import { createMockIntegrationLogger } from './logger';
@@ -21,9 +25,33 @@ export function createMockExecutionContext({
   instanceConfig,
 }: CreateMockExecutionContextOptions = {}): IntegrationExecutionContext {
   const logger = createMockIntegrationLogger();
+
   // copy local instance properties so that tests cannot
   // mutate the original object and cause unpredicable behavior
-  const instance = { ...LOCAL_INTEGRATION_INSTANCE, config: instanceConfig };
+  const instance = {
+    ...LOCAL_INTEGRATION_INSTANCE,
+  };
+
+  if (instanceConfig) {
+    instance.config = instanceConfig;
+  } else {
+    const configFields = loadInstanceConfigFields();
+    if (configFields) {
+      try {
+        instance.config = loadConfigFromEnvironmentVariables(configFields);
+      } catch (err) {
+        // failed to load configuration, not end of the world
+        // because this is only used in testing
+        //
+        // For convenience, we will generate a config for the developer
+        //
+        // this would generally only happen when a developer does not
+        // have an .env file configured or when an integration's test suite
+        // runs in CI
+        instance.config = generateInstanceConfig(configFields);
+      }
+    }
+  }
 
   return {
     logger,
@@ -46,4 +74,28 @@ export function createMockStepExecutionContext(
     ...createMockExecutionContext(options),
     jobState: createMockJobState(options),
   };
+}
+
+function generateInstanceConfig(
+  configFields: IntegrationInstanceConfigFieldMap,
+): IntegrationInstance['config'] {
+  return Object.entries(configFields).reduce(
+    (acc: IntegrationInstance['config'], [field, config]) => {
+      acc[field] = getInstanceConfigValueFromType(config);
+      return acc;
+    },
+    {},
+  );
+}
+
+function getInstanceConfigValueFromType(
+  config: IntegrationInstanceConfigField,
+) {
+  switch (config.type) {
+    case 'boolean':
+      return true;
+    case 'string':
+    default:
+      return 'STRING_VALUE';
+  }
 }